+++ EssentialFeed/Feed Cache/FeedCachePolicy.swift
@@ -1,0 +1,26 @@
+//
+//  FeedCachePolicy.swift
+//  EssentialFeed
+//
+//  Created by Ana Anguiano Cruz on 19/11/21.
+//
+
+import Foundation
+
+private final class FeedCachePolicy {
+    
+    private init() {}
+    
+    private static let calendar = Calendar(identifier: .gregorian)
+    
+    private static var maxCacheAgeInDays: Int {
+        return 7
+    }
+    
+    static func validate(_ timestamp: Date, against date: Date) -> Bool {
+        guard let maxCacheAge = calendar.date(byAdding: .day, value: maxCacheAgeInDays, to: timestamp) else {
+            return false
+        }
+        return date < maxCacheAge
+    }
+}
